move news-card-list
